fix: incorrect sourcemap (#457)
* fix: incorrect sourcemap

* fix: can not require esm module

* fix: not compile user's transform

* chore: remove require.resolve dir
