+++ packages/webpack-config/src/getCompilerPlugins.ts
@@ -3,18 +3,10 @@ import type { BuildOptions } from 'esbuild';
 import { createUnplugin } from 'unplugin';
 import type { UnpluginOptions } from 'unplugin';
 import compilationPlugin from './unPlugins/compilation.js';
+import compileExcludes from './compileExcludes.js';
 import type { WebpackConfig } from './index.js';
 
 type Compiler = 'webpack' | 'esbuild';
-
-const SKIP_COMPILE = [
-  // polyfill and helpers
-  'core-js', 'core-js-pure', '@swc/helpers', '@babel/runtime',
-  // built-in runtime
-  'react', 'react-dom',
-  // dev dependencies
-  '@pmmmwh/react-refresh-webpack-plugin', 'webpack', 'webpack-dev-server', 'react-refresh',
-];
 
 function getPluginTransform(plugin: UnpluginOptions, type: 'esbuild' | 'webpack') {
   const { transform } = plugin;
@@ -46,17 +38,15 @@ function getCompilerPlugins(config: Config, compiler: Compiler) {
     cacheDir,
   } = config;
   const compilerPlugins = [];
-  const compileExcludes = [
-    new RegExp(SKIP_COMPILE.map((dep) => `node_modules/?.+${dep}/`).join('|')),
-    /bundles\/compiled/,
-  ];
+
   // Add custom transform before swc compilation so the source code can be got before transformed.
   compilerPlugins.push(
     ...(transformPlugins.filter(({ enforce }) => !enforce || enforce === 'pre') || []),
     ...transforms.map((transform, index) => ({ name: `transform_${index}`, transform })),
   );
-
-  if (swcOptions) {
+  // Use webpack loader instead of webpack plugin to do the compilation.
+  // Reason: https://github.com/unjs/unplugin/issues/154
+  if (swcOptions && compiler !== 'webpack') {
     compilerPlugins.push(compilationPlugin({
       cacheDir,
       sourceMap,
