+++ packages/webpack-config/src/index.ts
@@ -21,6 +21,8 @@ import configCss from './config/css.js';
 import AssetsManifestPlugin from './webpackPlugins/AssetsManifestPlugin.js';
 import getCompilerPlugins from './getCompilerPlugins.js';
 import getSplitChunksConfig, { FRAMEWORK_BUNDLES } from './config/splitChunks.js';
+import compilationPlugin from './unPlugins/compilation.js';
+import compileExcludes from './compileExcludes.js';
 
 const require = createRequire(import.meta.url);
 const { merge } = lodash;
@@ -85,6 +87,8 @@ const getWebpackConfig: GetWebpackConfig = ({ rootDir, config, webpack, runtimeT
     devServer,
     fastRefresh,
     logging,
+    swcOptions,
+    compileIncludes,
   } = config;
   const absoluteOutputDir = path.isAbsolute(outputDir) ? outputDir : path.join(rootDir, outputDir);
   const dev = mode !== 'production';
@@ -147,7 +151,15 @@ const getWebpackConfig: GetWebpackConfig = ({ rootDir, config, webpack, runtimeT
       ascii_only: true,
     },
   }, minimizerOptions);
-
+  const compilation = compilationPlugin({
+    cacheDir,
+    sourceMap,
+    fastRefresh,
+    mode,
+    compileIncludes,
+    compileExcludes,
+    swcOptions,
+  });
   const webpackConfig: WebpackConfig = {
     mode,
     experiments: {
@@ -174,6 +186,17 @@ const getWebpackConfig: GetWebpackConfig = ({ rootDir, config, webpack, runtimeT
         },
       },
       rules: [
+        // Use webpack loader instead of webpack plugin which is generated by unplugin.
+        // Reason: https://github.com/unjs/unplugin/issues/154
+        {
+          test: compilation.transformInclude,
+          use: {
+            loader: require.resolve('./webpackLoaders/compilationLoader.cjs'),
+            options: {
+              transform: compilation.transform,
+            },
+          },
+        },
         ...loaders,
       ],
     },
